Fixed LootGenerator exceptions adding dragonscales
Changed to create loot using an existing ItemTemplate from the DB. Creating a new one causes the server to try to add the ItemTemplate to the ItemTemplate table, causing an exception which then mucks up the player's inventory.  This only happens if there isn't already an item of that type in the player's inventory, otherwise it just increases the count instead.