Redid LootGeneratorDragonscales
* Fixed formatting

Pesky tabs!

* Reduced drop count & considers target

The previous implementation appeared to be a copy/paste of the loot generator for Aurulite or Atlantean Glass.  As a result, it was dropping large stacks of dragon scales when it shouldn't be.

This redo checks for named dragons and drops the correct number of scales for named dragons while still allowing amounts dropped to be determined by server properties, so it can be used for epic and other named dragons rather than using loot templates.

* Changed LOOTGENERATOR_DRAGONSCALES_NAMED_COUNT to INT

Since LootGeneratorDragonscales isn't dropping huge bundles of scales any more, decimal values for LOOTGENERATOR_DRAGONSCALES_NAMED_COUNT don't make sense.  Changing it to an int will cause an error in servers using a decimal value and let the server admin know to change it, rather than having it round or drop decimals without the admin knowing there's a problem at all.

* Removed copy/paste error