Text correction in Merchant command + CheckPlayerAggro() respecting Z-axis
This small change to StandardMobBrain.CheckPlayerAggro() ensures that the Z-axis is respected in dungeons analogue to how it is done in CheckNpcAggro(). (#295)

Mobs in dungeons will no longer aggro on players in rooms far above or below them.
In Darkness falls in particular this would cause body pulls of high level mobs appearing seemingly out of thin air or coming out of walls.